gcp/requirements.txt
Flask
flask-restful
google-cloud-datastore